+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## dev
-* relax rules for guessing fixtures in decorator steps, introduce new config option `statefulPoms` for more strict checks ([#102](https://github.com/vitalets/playwright-bdd/issues/102))
+* feature: add `@slow` special tag ([#138](https://github.com/vitalets/playwright-bdd/issues/138))
+* improvement: relax rules for guessing fixtures in decorator steps, introduce new config option `statefulPoms` for more strict checks ([#102](https://github.com/vitalets/playwright-bdd/issues/102))
 
 ## 6.2.0
 * feature: add `--version` cli flag ([#134](https://github.com/vitalets/playwright-bdd/issues/134))